Human_we_will_find_and_we_will_win_W11_homework_end

+++ src/main/java/com/example/demo/controller/BlogController.java
@@ -89,7 +89,7 @@ public class BlogController {
     //     return "redirect:/article_list";
     // }
     @GetMapping("/board_view/{id}") // 게시판 링크 지정
-    public String board_view(Model model, @PathVariable Long id ) { // , @RequestParam int orderNumber 삭제 오류남
+    public String board_view(Model model, @PathVariable Long id, HttpSession session) { // , @RequestParam int orderNumber 삭제 오류남
         Optional<Board> list = blogService.findById(id); // 선택한 게시판 글
         if (list.isPresent()) {
             model.addAttribute("boards", list.get()); // 존재할 경우 실제 Board 객체를 모델에 추가
@@ -98,6 +98,9 @@ public class BlogController {
             // 처리할 로직 추가 (예: 오류 페이지로 리다이렉트, 예외 처리 등)
             return "/error_page/article_error"; // 오류 처리 페이지로 연결
         }
+        String email = (String) session.getAttribute("email"); // 세션에서 이름 확인
+        model.addAttribute("email", email); // 로그인 사용자(이메일)
+
         return "board_view"; // .HTML 연결
     }
     @GetMapping("/board_edit/{id}") // 게시판링크지정
@@ -112,12 +115,27 @@ public class BlogController {
         return "board_edit"; // .HTML 연결
     }
     @PutMapping("api/board_edit/{id}")
-    public String updataBoard(@PathVariable Long id, @ModelAttribute AddArticleRequest request) {
+    public String updataBoard(@PathVariable Long id, @ModelAttribute AddArticleRequest request, HttpSession session) {
+        String email = (String) session.getAttribute("email"); // 세션에서 이름 확인
+        Optional<Board> list = blogService.findById(id);
+        Board board = list.get(); // 여기서 실제 Board 꺼내기
+
+        if (!board.getUser().equals(email)) {
+            return "redirect:/board_list?error=unauthorized";
+        }
+
         blogService.update(id, request);
         return "redirect:/board_list"; // 글 수정 이후 .html 연결
     }
     @DeleteMapping("/api/board_delete/{id}")
-    public String deleteBoard(@PathVariable Long id) {
+    public String deleteBoard(@PathVariable Long id, HttpSession session) {
+        String email = (String) session.getAttribute("email"); // 세션에서 이름 확인
+        Optional<Board> list = blogService.findById(id);
+        Board board = list.get(); // 여기서 실제 Board 꺼내기
+        if (!board.getUser().equals(email)) {
+            return "redirect:/board_list?error=unauthorized";            
+        }
+
         blogService.delete(id);
         return "redirect:/board_list";
     }
@@ -127,7 +145,10 @@ public class BlogController {
     }
 
     @PostMapping("/api/boards")
-    public String addboards(@ModelAttribute AddArticleRequest request) {
+    public String addboards(@ModelAttribute AddArticleRequest request, HttpSession session) {
+        String email = (String) session.getAttribute("email"); // 세션에서 이름 확인
+        System.out.println("작성자: " + email);
+        request.setUser(email);  // 또는 getNickname()
         blogService.save(request); // 게시글저장
         return "redirect:/board_list";
     }
@@ -149,11 +170,11 @@ public class BlogController {
         }
         int startNumber = (page * 5) + 1; // 페이지 시작 번호 계산
         model.addAttribute("startNumber", startNumber); // 모델에 추가
-         model.addAttribute("email", email); // 로그인 사용자(이메일)
+        model.addAttribute("email", email); // 로그인 사용자(이메일)
         model.addAttribute("boards", list); // 모델에 추가
         model.addAttribute("totalPages", list.getTotalPages()); // 페이지 크기
         model.addAttribute("currentPage", page); // 페이지 번호
         model.addAttribute("keyword", keyword); // 키워드
         return "board_list"; // .HTML 연결
     }
-    }
+}
